esta arreglado el asunto

+++ carro/models.py
@@ -111,8 +111,9 @@ class LineaCarro(models.Model):
 			except ObjectDoesNotExist:
 				super(LineaCarro, self).save(*args, **kwargs)
 			else:
-				coincidencias.delete()
-				self.cantidad = coincidencias.cantidad+self.cantidad
+				if self.pk!=coincidencias.pk:
+					coincidencias.delete()
+					self.cantidad = coincidencias.cantidad+self.cantidad
 				super(LineaCarro, self).save(*args, **kwargs)
 
 class Prueba(models.Model):
